azure-services-image-analysis/requirements.txt
dotenv
azure-ai-vision-imageanalysis==1.0.0b1
pillow
matplotlib